refactor : 픽스처 클래스 엔티티와 동일한 패키지로 이동
