+++ src/test/java/com/gwangya/purchase/domain/PurchaseInfoFixture.java
@@ -1,11 +1,10 @@
-package com.gwangya.fixture.purchase;
+package com.gwangya.purchase.domain;
 
 import static com.gwangya.purchase.domain.ReceivingType.*;
 
 import java.time.LocalDateTime;
 
 import com.gwangya.performance.domain.PerformanceDetail;
-import com.gwangya.purchase.domain.PurchaseInfo;
 import com.gwangya.purchase.domain.vo.PurchaseSerialNumber;
 import com.gwangya.user.domain.User;
